% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_boot_two_groups.R
\name{boot_two_groups}
\alias{boot_two_groups}
\title{Bootstrap sampling for difference in means between two groups}
\usage{
boot_two_groups(
  data,
  groups_col,
  n_max,
  n_min = 3,
  iter = 29,
  response,
  group1,
  group2
)
}
\arguments{
\item{data}{Data frame contain raw data.}

\item{groups_col}{Factor. Column containing names of two populations to compare}

\item{n_max}{Numeric. Maximum sample size to extrapolate simulations.}

\item{n_min}{Numeric. Minimum sample size to extrapolate simulations. Defaults to 3.}

\item{iter}{Numeric. Number of bootstrap samples to draw. Defaults to 29.}

\item{response}{Numeric. Column containing thermal limit data for individual samples.}

\item{group1}{String. Name of first population to compare.}

\item{group2}{String. Name of second population to compare.}
}
\value{
A data frame of bootstrap resamples
}
\description{
Calculate difference in mean CT limits between two groups.
}
\examples{
head(coreid_data)
sims <- boot_two_groups(data = coreid_data,
                        groups_col = col,
                        response = response,
                        group1 = "Catorhintha schaffneri_APM",
                        group2 = "Catorhintha schaffneri_NPM",
                        n_max = 49,
                        iter = 99)
}
