% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calc_ci_coverage.R
\name{calc_ci_coverage}
\alias{calc_ci_coverage}
\title{Calculate coverage of confidence intervals}
\usage{
calc_ci_coverage(x, groups_col)
}
\arguments{
\item{x}{Data frame. Output of boot_sample function.}

\item{groups_col}{Factor. Column containing name(s) of population(s) of interest}
}
\value{
A data frame of CTL summary statistics from bootstrap resamples
}
\description{
Calculate how many bootstrap sample CTL estimates fall within bounds of 95%
confidence interval.
}
\examples{
head(coreid_data)
# First draw bootstrap samples
sims <- boot_sample(data = coreid_data,
                    groups_col = col,
                    response = response,
                    n_max = 49,
                    iter = 99)
# Now we can calculate coverage
cover <- calc_ci_coverage(x= sims,
                          groups_col = col)
}
