% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreid_data.R
\docType{data}
\name{coreid_data}
\alias{coreid_data}
\title{Example critical thermal limit data for ThermalSampleN package}
\format{
A data frame with 60 rows and 2 variables:
\describe{
  \item{col}{insect population tested, string}
  \item{response}{CTmin value for individual insect, in degrees celsius}
  ...
}
}
\source{
"Unpublished data provided by Philippa Muskett
(Centre for Biological Control, Rhodes University, South Africa)
}
\usage{
coreid_data
}
\description{
A dataset containing critical thermal limit (CTmin) data for the
sap-sucking bug, Catorintha schaffneri.
}
\keyword{datasets}
