% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_tost.R
\name{equiv_tost}
\alias{equiv_tost}
\title{equiv_tost}
\arguments{
\item{data}{Data frame contains raw data. Must contain a column with a population identifier (e.g. population ID), and a column containing critical
thermal limit data (e.g. temperatures at which critical limits are reached).}

\item{groups_col}{Factor. Column containing the name of the population of interest (group ID)}

\item{groups_which}{Character. Which population should be analysed?}

\item{response}{Numeric. Column containing thermal limit data for individual samples}

\item{skews}{Numeric vector containing skewness parameter(s)}

\item{equiv_margin}{Equivalence of subsets to full population CT estimate (unit = degree Celcius)}

\item{pop_n}{Size of population to sample (will test subsamples of size pop_n - x against pop_n for equivalence) Defaults to population size = 30}

\item{colrs}{Colours of the skewness paramaters}
}
\value{
Two plots; (a) equivalence of means, and (b) equivalence of variances
}
\description{
Perform a Test of Total Equivalence as per Duffy et al. (2021)
}
\examples{

head(coreid_data)
res <- equiv_tost(data = coreid_data,
                    groups_col = col,
                    groups_which = "Catorhintha schaffneri_APM",
                    response = response,
                    skews = c(1,10),
                    colrs = c("lightblue", "lightpink"),
                    equiv_margin = 1,
                    pop_n = 5)
}
