% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_one_group.R
\name{plot_one_group}
\alias{plot_one_group}
\title{Plot output from boot_sample}
\usage{
plot_one_group(x = sims, n_min = 3, n_max)
}
\arguments{
\item{x}{Output from boot_sample function. Defaults to 'sims'.}

\item{n_min}{Numeric. Minimum sample size to extrapolate simulations.
Defaults to 3.}

\item{n_max}{Numeric. Maximum sample size to extrapolate simulations.}
}
\description{
Plot output from boot_sample.
}
\examples{
sims <- boot_sample(coreid_data,
                    groups_col = col,
                    groups_which = "Catorhintha schaffneri_APM",
                    n_max = 30,
                    response = response)
plot_one_group(x = sims,
               n_min = 3,
               n_max = 15)

}
