% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_two_groups.R
\name{plot_two_groups}
\alias{plot_two_groups}
\title{Plot output from boot_two_groups}
\usage{
plot_two_groups(x = sims, n_min = 3, n_max)
}
\arguments{
\item{x}{Output from boot_two_groups function. Defaults to 'sims'.}

\item{n_min}{Numeric. Minimum sample size to extrapolate simulations.
Defaults to 3.}

\item{n_max}{Numeric. Maximum sample size to extrapolate simulations.}
}
\description{
Plot output from boot_two_groups.
}
\examples{
sims <- boot_two_groups(data = coreid_data,
                        groups_col = col,
                        response = response,
                        group1 = "Catorhintha schaffneri_APM",
                        group2 = "Catorhintha schaffneri_NPM",
                        n_max = 49,
                        iter = 99)
plots <- plot_two_groups(x = sims,
                         n_min = 3,
                         n_max = 49)

}
