% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_boot_sample.R
\name{boot_sample}
\alias{boot_sample}
\title{Bootstrap sampling to calculate summary statistics of CTL values}
\usage{
boot_sample(
  data,
  groups_col,
  groups_which,
  n_max,
  n_min = 3,
  iter = 29,
  response
)
}
\arguments{
\item{data}{Data frame contain raw data. Must contain a column with a population
identifier (e.g. population ID), and a column containin critical
thermal limit data (e.g. temperatures at which critical limits are reached).}

\item{groups_col}{Factor. Column containing name(s) of population(s) of interest}

\item{groups_which}{Character. Which population should be analysed?}

\item{n_max}{Numeric. Maximum sample size to extrapolate simulations.}

\item{n_min}{Numeric. Minimum sample size to extrapolate simulations. Defaults to 3.}

\item{iter}{Numeric. Number of bootstrap samples to draw. Defaults to 29.}

\item{response}{Numeric. Column containing thermal limit data for individual samples}
}
\value{
A data frame of CTL summary statistics from bootstrap resamples
}
\description{
Calculate mean and CI's of CTL for a single population
}
\examples{
head(coreid_data)
sims <- boot_sample(data = coreid_data,
                    groups_col = col,
                    groups_which = "Catorhintha schaffneri_APM",
                    response = response,
                    n_max = 49,
                    iter = 99)
}
